completed 8 points

+++ src/main/liquibase/sql/03.create-book_author-table.sql
@@ -1,4 +1,4 @@
 create table book_author(
-author_id int references authors(id),
-book_id int references  books(id)
+author_id int8 references authors(id),
+book_id int8 references  books(id)
 );
