+++ src/main/liquibase/sql/09.create-users-table.sql
@@ -1,7 +1,7 @@
 create table users
 (
-    id            serial primary key,
-    credential_id int references credentials (id),
+    id            bigserial primary key,
+    credential_id int8 references credentials (id),
     first_name    varchar(64),
     surname       varchar(64),
     telephone     varchar(64)
